backend tests pass \o/

+++ backend/src/resolvers.ts
@@ -1,38 +1,56 @@
 import { GraphQLJSON } from 'graphql-type-json'
-import { RegistrationStatus } from './generated/graphql.js'
-
 import {
-  QueryResolvers,
-  MutationResolvers,
-  UserResolvers,
-  TableResolvers,
   ExhibitorResolvers,
   ExhibitResolvers,
+  MutationResolvers,
+  QueryResolvers,
   RegistrationResolvers,
+  RegistrationStatus,
+  TableResolvers,
+  UserResolvers,
 } from './generated/graphql.js'
 import { Context } from './app/context.js'
 import { wrap } from '@mikro-orm/core'
 import { GraphQLError } from 'graphql/error/index.js'
+import { requireAdmin } from './db.js'
+import { Exhibit } from './entities.js'
 
 const queryResolvers: QueryResolvers<Context> = {
-  getUser: async (_, { id }, { db }) => db.user.findOneOrFail({ id }),
-  getUsers: async (_, _args, { db }) => db.user.findAll(),
+  getUser: async (_, { id }, { db, user }) => {
+    requireAdmin(user)
+    return db.user.findOneOrFail({ id })
+  },
+  getUsers: async (_, _args, { db, user }) => {
+    requireAdmin(user)
+    return db.user.findAll()
+  },
   getCurrentUser: async (_, _args, { user }) => user,
   getUserByEmail: async (_, { email }, { db }) =>
     db.user.findOneOrFail({ email }),
+  // @ts-expect-error ts2345
   getTable: async (_, { number }, { db }) => db.table.findOneOrFail({ number }),
+  // @ts-expect-error ts2345
   getTables: async (_, _args, { db }) => db.table.findAll(),
+  // @ts-expect-error ts2345
   getExhibitor: async (_, { id }, { db }) =>
     await db.exhibitor.findOneOrFail({ id }),
+  // @ts-expect-error ts2345
   getExhibitors: async (_, _args, { db }) => db.exhibitor.findAll(),
+  // @ts-expect-error ts2345
   getExhibit: async (_, { id }, { db }) => db.exhibit.findOneOrFail({ id }),
+  // @ts-expect-error ts2345
   getExhibits: async (_, _args, { db }) => db.exhibit.findAll(),
   getExhibition: async (_, { id }, { db }) =>
     db.exhibition.findOneOrFail({ id }),
   getExhibitions: async (_, _args, { db }) => db.exhibition.findAll(),
-  getRegistration: async (_, { id }, { db }) =>
-    db.registration.findOneOrFail({ id }),
-  getRegistrations: async (_, _args, { db }) => db.registration.findAll(),
+  getRegistration: async (_, { id }, { db, user }) => {
+    requireAdmin(user)
+    return db.registration.findOneOrFail({ id })
+  },
+  getRegistrations: async (_, _args, { db, user }) => {
+    requireAdmin(user)
+    return db.registration.findAll()
+  },
 }
 
 const mutationResolvers: MutationResolvers<Context> = {
@@ -57,28 +75,35 @@ const mutationResolvers: MutationResolvers<Context> = {
     await db.user.resetPassword(token, password)
     return true
   },
-  updateUserProfile: async (_, { input }, { user, db }) => {
+  updateUserProfile: async (_, { input }, { user }) => {
     if (!user) {
       throw new Error('You must be logged in to update your profile')
     }
     wrap(user).assign(input)
     return user
   },
-  claimTable: async (_, { number }, { db, exhibitor, user }) => {
+  // @ts-expect-error ts2345
+  claimTable: async (_, { number }, { db, exhibitor }) => {
     if (!exhibitor) {
       throw new Error('You must be logged in to claim a table')
     }
     return await db.table.claim(number, exhibitor)
   },
+  // @ts-expect-error ts2345
   releaseTable: async (_, { number }, { db, exhibitor, user }) => {
-    await db.table.release(number, user?.isAdministrator ? null : exhibitor)
+    return await db.table.release(
+      number,
+      user?.isAdministrator ? null : exhibitor,
+    )
   },
+  // @ts-expect-error ts2345
   assignTable: async (_, { number, exhibitorId }, { db, exhibition }) => {
     const exhibitor = await db.exhibitor.findOneOrFail({ id: exhibitorId })
     const table = await db.table.findOneOrFail({ exhibition, number })
     table.exhibitor = exhibitor
     return table
   },
+  // @ts-expect-error ts2345
   createExhibit: async (
     _,
     { title, text, table },
@@ -87,51 +112,71 @@ const mutationResolvers: MutationResolvers<Context> = {
     if (!exhibitor) {
       throw new Error('You must be logged in to create an exhibit')
     }
-    return db.exhibit.create({
+    const exhibit = db.em.getRepository(Exhibit).create({
       exhibition,
       title,
       text,
       table,
       exhibitor,
     })
+    await db.em.persist(exhibit).flush()
+    return exhibit
   },
-  updateExhibit: async (_, { id, title, text, table }, { db }) => {
+  // @ts-expect-error ts2345
+  updateExhibit: async (_, { id, ...rest }, { db, exhibitor, user }) => {
     const exhibit = await db.exhibit.findOneOrFail({ id })
-    return wrap(exhibit).assign({ title, text, table })
+    if (exhibitor !== exhibit.exhibitor && !user?.isAdministrator) {
+      throw new Error('You do not have permission to update this exhibit')
+    }
+    return wrap(exhibit).assign(rest)
   },
-  createRegistration: async (
-    _,
-    { name, email, nickname, topic, message, data },
-    { exhibition, db },
-  ) =>
-    db.registration.create({
-      status: RegistrationStatus.New,
+  register: async (_, { input }, { exhibition, db }) => {
+    const { email, message, ...rest } = input
+    const existing = await db.registration.findOne({
+      email: email,
       exhibition,
-      name,
+    })
+    if (existing) {
+      throw new GraphQLError('The email address is already registered')
+    }
+    return await db.registration.register({
+      exhibition,
+      status: RegistrationStatus.New,
+      message: message || undefined,
       email,
-      nickname,
-      topic,
-      message,
-      data,
-    }),
-  updateRegistration: async (_, { id, status, notes }, { db }) => {
+      ...rest,
+    })
+  },
+  updateRegistrationNotes: async (_, { id, notes }, { db, user }) => {
+    requireAdmin(user)
     const registration = await db.registration.findOneOrFail({ id })
-    wrap(registration).assign({ status, notes })
+    wrap(registration).assign({ notes })
     return registration
   },
-  approveRegistration: async (_, { id, siteUrl }, { db }) => {
+  approveRegistration: async (_, { id, siteUrl }, { db, user }) => {
+    requireAdmin(user)
     const registration = await db.registration.findOneOrFail({ id })
     await db.registration.approve(registration, siteUrl)
     return true
   },
-  rejectRegistration: async (_, { id }, { db }) => {
+  rejectRegistration: async (_, { id }, { db, user }) => {
+    requireAdmin(user)
     const registration = await db.registration.findOneOrFail({ id })
     await db.registration.reject(registration)
     return true
   },
-  deleteRegistration: async (_, { id }, { db }) => {
+  deleteRegistration: async (_, { id }, { db, user }) => {
+    requireAdmin(user)
     const registration = await db.registration.findOneOrFail({ id })
+    if (registration.status === RegistrationStatus.Approved) {
+      throw new Error('Cannot delete approved registration')
+    }
     db.em.remove(registration)
+    return true
+  },
+  setRegistrationInProgress: async (_, { id }, { db }) => {
+    const registration = await db.registration.findOneOrFail({ id })
+    await db.registration.inProgress(registration)
     return true
   },
 }
@@ -145,19 +190,25 @@ const tableResolvers: TableResolvers = {
     table.exhibitor
       ? db.exhibitor.findOneOrFail({ id: table.exhibitor.id })
       : null,
+  exhibits: async (table, _, { db }) => db.exhibit.find({ table: table }),
 }
 
 const exhibitorResolvers: ExhibitorResolvers = {
   user: async (exhibitor, _, { db }) =>
     db.user.findOneOrFail({ id: exhibitor.user.id }),
-  exhibits: async (exhibitor) => exhibitor.exhibits || [],
-  tables: async (exhibitor) => exhibitor.tables || [],
+  exhibits: async (exhibitor, _, { db }) =>
+    db.exhibit.find({
+      id: { $in: exhibitor.exhibits?.map((e) => e.id) || [] },
+    }),
+  tables: async (exhibitor, _, { db }) =>
+    db.table.find({ id: { $in: exhibitor.tables?.map((t) => t.id) || [] } }),
 }
 
 const exhibitResolvers: ExhibitResolvers = {
-  exhibitor: async (exhibit) => {
-    return exhibit.exhibitor
-  },
+  exhibitor: async (exhibit, _, { db }) =>
+    db.exhibitor.findOneOrFail({ id: exhibit.exhibitor.id }),
+  table: async (exhibit, _, { db }) =>
+    exhibit.table ? db.table.findOneOrFail({ id: exhibit.table.id }) : null,
 }
 
 const registrationResolvers: RegistrationResolvers = {
